arxiv_github_agent/requirements.txt
requests
python-dateutil
python-dotenv
arxiv
PyGithub
